JAMES-3440 Type limit as an Int

+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/core/Query.scala
@@ -42,7 +42,7 @@ object Position {
   }
 }
 
-case class LimitUnparsed(value: Long) extends AnyVal
+case class LimitUnparsed(value: Int) extends AnyVal
 
 object Limit {
   type Limit = Int Refined Positive
